refactor: `config` -> `model`
I am already getting confused by all the multiple levels of configuration so it's better for general readability to name the original `config` file `model`, since it basically provides the data model
